+++ RxSwift/Observables/Implementations/ShareReplay1.swift
@@ -12,7 +12,8 @@ final class ShareReplay1<Element>
     , ObserverType
     , SynchronizedUnsubscribeType {
 
-    typealias DisposeKey = Bag<AnyObserver<Element>>.KeyType
+    typealias Observers = AnyObserver<Element>.s
+    typealias DisposeKey = Observers.KeyType
 
     private let _source: Observable<Element>
 
@@ -22,7 +23,7 @@ final class ShareReplay1<Element>
     private var _element: Element?
     private var _stopped = false
     private var _stopEvent = nil as Event<Element>?
-    private var _observers = Bag<(Event<Element>) -> ()>()
+    private var _observers = Observers()
 
     init(source: Observable<Element>) {
         self._source = source
@@ -81,10 +82,10 @@ final class ShareReplay1<Element>
         dispatch(_synchronized_on(event), event)
     }
 
-    func _synchronized_on(_ event: Event<E>) -> Bag<(Event<Element>) -> ()> {
+    func _synchronized_on(_ event: Event<E>) -> Observers {
         _lock.lock(); defer { _lock.unlock() }
         if _stopped {
-            return Bag()
+            return Observers()
         }
 
         switch event {