+++ RxSwift/Observables/Implementations/ShareReplay1WhileConnected.swift
@@ -10,7 +10,8 @@ fileprivate final class ShareReplay1WhileConnectedConnection<Element>
     : ObserverType
     , SynchronizedUnsubscribeType {
     typealias E = Element
-    typealias DisposeKey = Bag<(Event<Element>) -> ()>.KeyType
+    typealias Observers = AnyObserver<Element>.s
+    typealias DisposeKey = Observers.KeyType
 
     typealias Parent = ShareReplay1WhileConnected<Element>
     private let _parent: Parent
@@ -18,7 +19,7 @@ fileprivate final class ShareReplay1WhileConnectedConnection<Element>
 
     private let _lock: RecursiveLock
     private var _disposed: Bool = false
-    fileprivate var _observers = Bag<(Event<Element>) -> ()>()
+    fileprivate var _observers = Observers()
     fileprivate var _element: Element?
 
     init(parent: Parent, lock: RecursiveLock) {
@@ -33,9 +34,9 @@ fileprivate final class ShareReplay1WhileConnectedConnection<Element>
         dispatch(observers, event)
     }
 
-    final private func _synchronized_on(_ event: Event<E>) -> Bag<(Event<Element>) -> ()> {
+    final private func _synchronized_on(_ event: Event<E>) -> Observers {
         if _disposed {
-            return Bag()
+            return Observers()
         }
 
         switch event {
@@ -69,7 +70,7 @@ fileprivate final class ShareReplay1WhileConnectedConnection<Element>
         if _parent._connection === self {
             _parent._connection = nil
         }
-        _observers = Bag()
+        _observers = Observers()
         _subscription.dispose()
     }